requirements.txt
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
requests==2.28.1
python-dotenv==1.0.0